Change in adjustTimeZone
3600 changed to 3600.0 and casted result into unsigned long

+++ NTPtimeESP.cpp
@@ -130,7 +130,7 @@ boolean NTPtime::daylightSavingTime(unsigned long _timeStamp) {
 
 unsigned long NTPtime::adjustTimeZone(unsigned long _timeStamp, float _timeZone, byte _DayLightSaving) {
   strDateTime _tempDateTime;
-  _timeStamp += _timeZone *  3600; // adjust timezone
+  _timeStamp += (unsigned long)(_timeZone *  3600.0); // adjust timezone
   if (_DayLightSaving ==1 && summerTime(_timeStamp)) _timeStamp += 3600; // European Summer time
   if (_DayLightSaving ==2 && daylightSavingTime(_timeStamp)) _timeStamp += 3600; // US daylight time
   return _timeStamp;
